更新JQLibrary

+++ lib/JQLibrary/src/JQQRCodeWriter/JQQRCodeWriter.cpp
@@ -69,7 +69,7 @@ QImage JQQRCodeWriter::makeQRcode(const QString &data, const QSize &size)
 
     for ( int y = 0; y < s; ++y )
     {
-        const int &&yy = y * s;
+        const int &&yy = static_cast< int >( y * s );
         for( int x = 0; x < s; ++x )
         {
             const int &&xx = yy + x;
